# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Recolección de Residuos",
        "description": "Documentación de la API para el sistema de gestión de rutas de recolección.",
        "contact": {
            "email": "soporte@tuempresa.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/recorridos/{recorrido}/posiciones": {
            "get": {
                "tags": [
                    "Posiciones"
                ],
                "summary": "Listar posiciones de un recorrido",
                "operationId": "d48be9ac413dcd1bb7645b6de7fa8308",
                "parameters": [
                    {
                        "name": "recorrido",
                        "in": "path",
                        "description": "ID del recorrido",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "perfil_id",
                        "in": "query",
                        "description": "UUID del perfil propietario para verificar el permiso.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de posiciones."
                    },
                    "403": {
                        "description": "Acción no autorizada."
                    },
                    "404": {
                        "description": "Recorrido no encontrado."
                    }
                }
            },
            "post": {
                "tags": [
                    "Posiciones"
                ],
                "summary": "Registrar una nueva posición",
                "operationId": "7bd4f3a61bc3502236f14343142aec81",
                "parameters": [
                    {
                        "name": "recorrido",
                        "in": "path",
                        "description": "ID del recorrido",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lat",
                                    "lon",
                                    "perfil_id"
                                ],
                                "properties": {
                                    "lat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 3.42158
                                    },
                                    "lon": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -76.5205
                                    },
                                    "perfil_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Posición registrada."
                    },
                    "403": {
                        "description": "Acción no autorizada."
                    },
                    "422": {
                        "description": "Validación fallida."
                    }
                }
            }
        },
        "/api/recorridos": {
            "get": {
                "tags": [
                    "Recorridos"
                ],
                "summary": "Listar recorridos por perfil",
                "operationId": "3679a4c043b8891beea6689d807d4f11",
                "parameters": [
                    {
                        "name": "perfil_id",
                        "in": "query",
                        "description": "UUID del perfil para filtrar los recorridos.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de recorridos."
                    },
                    "422": {
                        "description": "Error de validación."
                    }
                }
            },
            "post": {
                "tags": [
                    "Recorridos"
                ],
                "summary": "Iniciar un nuevo recorrido",
                "description": "Crea un nuevo registro de recorrido asociado a un perfil.",
                "operationId": "b4740bebd0b2ff5cd153b7b32011ef10",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ruta_id",
                                    "vehiculo_id",
                                    "perfil_id"
                                ],
                                "properties": {
                                    "ruta_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "vehiculo_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "perfil_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Recorrido iniciado exitosamente."
                    },
                    "422": {
                        "description": "Datos de validación inválidos."
                    }
                }
            }
        },
        "/api/recorridos/{recorrido}/finalizar": {
            "post": {
                "tags": [
                    "Recorridos"
                ],
                "summary": "Finalizar un recorrido existente",
                "operationId": "137a4cd897a71f875627eda1291cf3c3",
                "parameters": [
                    {
                        "name": "recorrido",
                        "in": "path",
                        "description": "ID del recorrido a finalizar.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "perfil_id"
                                ],
                                "properties": {
                                    "perfil_id": {
                                        "description": "ID del perfil propietario del recorrido.",
                                        "type": "string",
                                        "format": "uuid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recorrido finalizado."
                    },
                    "403": {
                        "description": "Acción no autorizada. El perfil no corresponde."
                    },
                    "404": {
                        "description": "Recorrido no encontrado."
                    }
                }
            }
        },
...
    },
...
}